# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TourFlow AI
  description: 'TourFlow AI API contract — generated from the FastAPI implementation
    (source of truth: backend/app). Regenerate: python -c "from app.main import app;
    ..."'
  version: 1.0.0
paths:
...
  /api/v1/trips/{trip_id}/bookings:
    post:
      tags:
      - bookings
      summary: Create Booking
      operationId: create_booking_api_v1_trips__trip_id__bookings_post
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - bookings
      summary: List Bookings
      operationId: list_bookings_api_v1_trips__trip_id__bookings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingOut'
                title: Response List Bookings Api V1 Trips  Trip Id  Bookings Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/bookings/{booking_id}:
    patch:
      tags:
      - bookings
      summary: Update Booking
      operationId: update_booking_api_v1_bookings__booking_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: booking_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Booking Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/bookings:
    get:
      tags:
      - bookings
      summary: All Bookings
      operationId: all_bookings_api_v1_bookings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BookingOut'
                type: array
                title: Response All Bookings Api V1 Bookings Get
      security:
      - HTTPBearer: []
  /api/v1/trips/{trip_id}/pricing:
    get:
      tags:
      - bookings
      summary: Get Pricing
      description: Backend-authoritative cost estimate. Frontend totals are never
        trusted.
      operationId: get_pricing_api_v1_trips__trip_id__pricing_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/vendors:
    get:
      tags:
      - catalog
      summary: List Vendors
      operationId: list_vendors_api_v1_vendors_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VendorOut'
                title: Response List Vendors Api V1 Vendors Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - catalog
      summary: Create Vendor
      operationId: create_vendor_api_v1_vendors_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/vendors/{vendor_id}:
    patch:
      tags:
      - catalog
      summary: Update Vendor
      operationId: update_vendor_api_v1_vendors__vendor_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: vendor_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Vendor Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/destinations:
    get:
      tags:
      - catalog
      summary: List Destinations
      operationId: list_destinations_api_v1_destinations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DestinationOut'
                title: Response List Destinations Api V1 Destinations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BookingCreate:
      properties:
        service_type:
          type: string
          maxLength: 60
          minLength: 2
          title: Service Type
        service_name:
          type: string
          maxLength: 250
          minLength: 2
          title: Service Name
        vendor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendor Id
        amount:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Amount
        currency:
          type: string
          maxLength: 3
          minLength: 3
          title: Currency
          default: INR
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - service_type
      - service_name
      title: BookingCreate
    BookingOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        trip_id:
          type: string
          format: uuid
          title: Trip Id
        traveler_id:
          type: string
          format: uuid
          title: Traveler Id
        vendor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendor Id
        service_type:
          type: string
          title: Service Type
        service_name:
          type: string
          title: Service Name
        reference_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Reference Code
        status:
          $ref: '#/components/schemas/BookingStatus'
        amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Amount
        currency:
          type: string
          title: Currency
        booked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Booked At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - id
      - trip_id
      - traveler_id
      - service_type
      - service_name
      - status
      - currency
      title: BookingOut
    BookingStatus:
      type: string
      enum:
      - PLANNED
      - PENDING
      - CONFIRMED
      - CANCELLED
      - REBOOKED
      - COMPLETED
      title: BookingStatus
    BookingUpdate:
      properties:
        status:
          anyOf:
          - $ref: '#/components/schemas/BookingStatus'
          - type: 'null'
        amount:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Amount
        reference_code:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Reference Code
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: BookingUpdate
...
    DestinationOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        country:
          type: string
          title: Country
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        tags:
          items:
            type: string
          type: array
          title: Tags
          default: []
        avg_daily_cost:
          anyOf:
          - type: number
          - type: 'null'
          title: Avg Daily Cost
        currency:
          type: string
          title: Currency
      type: object
      required:
      - id
      - name
      - country
      - currency
      title: DestinationOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PricingOut:
      properties:
        trip_id:
          type: string
          format: uuid
          title: Trip Id
        currency:
          type: string
          title: Currency
        transportation:
          type: number
          title: Transportation
        accommodation:
          type: number
          title: Accommodation
        activities:
          type: number
          title: Activities
        other:
          type: number
          title: Other
        base_cost:
          type: number
          title: Base Cost
        additional_cost:
          type: number
          title: Additional Cost
        discount:
          type: number
          title: Discount
        estimated_total:
          type: number
          title: Estimated Total
        budget:
          anyOf:
          - type: number
          - type: 'null'
          title: Budget
        within_budget:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Within Budget
      type: object
      required:
      - trip_id
      - currency
      - transportation
      - accommodation
      - activities
      - other
      - base_cost
      - additional_cost
      - discount
      - estimated_total
      title: PricingOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VendorIn:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 2
          title: Name
        service_type:
          type: string
          maxLength: 60
          title: Service Type
        contact_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Contact Email
        contact_phone:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Contact Phone
        city:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: City
        profile_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Profile Id
      type: object
      required:
      - name
      - service_type
      title: VendorIn
    VendorOut:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 2
          title: Name
        service_type:
          type: string
          maxLength: 60
          title: Service Type
        contact_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Contact Email
        contact_phone:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Contact Phone
        city:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: City
        profile_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Profile Id
        id:
          type: string
          format: uuid
          title: Id
        is_active:
          type: boolean
          title: Is Active
        rating:
          anyOf:
          - type: number
          - type: 'null'
          title: Rating
      type: object
      required:
      - name
      - service_type
      - id
      - is_active
      title: VendorOut
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer